+++ src/stores/transcript_store.js
@@ -1,6 +1,6 @@
 import { create } from 'zustand'
 import { log } from 'mentie'
-import { load_asr_model, transcribe_audio } from '../modules/asr/asr_client.js'
+import { cancel_asr_work, load_asr_model, transcribe_audio } from '../modules/asr/asr_client.js'
 import { prepare_audio_file } from '../modules/audio/decode_audio.js'
 import { get_selected_model_profile } from '../modules/model/model_cache.js'
 import { clear_share_with_transcript, load_share_with_transcript, mark_share_status, save_manual_audio_file } from '../modules/share/share_db.js'
@@ -182,6 +182,8 @@ export const use_transcript_store = create( ( set, get ) => ( {
 
     cancel_processing: async () => {
         const { share } = get()
+
+        cancel_asr_work()
 
         set( {
             active_run_id: null,